requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
amqp==5.2.0
async-timeout==4.0.3
attrs==23.2.0
awsebcli==3.20.10
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
blinker==1.8.2
botocore==1.31.85
celery==5.4.0
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
gunicorn==22.0.0
html5lib==1.1
huey==2.5.1
idna==3.7
importlib-metadata==4.13.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
kombu==5.4.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
multidict==6.0.5
multitasking==0.0.11
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pandas-datareader==0.10.0
pathspec==0.10.1
patsy==0.5.6
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rq==1.16.2
scikit-learn==1.5.1
scipy==1.13.1
semantic-version==2.8.5
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
statsmodels==0.14.2
termcolor==1.1.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==3.0.3
yarl==1.9.4
yfinance==0.2.41
zipp==3.19.2
